refactor (state) : step3 - state 패키지 리팩토링
