+++ src/test/java/bowling/domain/state/ReadyTest.java
@@ -3,6 +3,9 @@ package bowling.domain.state;
 import bowling.domian.frame.exception.InvalidScoreCalculateException;
 import bowling.domian.state.*;
 import bowling.domian.state.exception.InvalidPinCountException;
+import bowling.domian.state.finished.Strike;
+import bowling.domian.state.running.FirstBowl;
+import bowling.domian.state.running.Ready;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
